+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+- **Fixed:** GraphQL conditional import.
+
+<br>
+
 ## [v1.6.0-beta.1](https://github.com/vazco/uniforms/releases/tag/v1.6.0-beta.1)
 - **Changed:** Added GraphQL schema support. [\#118](https://github.com/vazco/uniforms/issues/118)
 - **Changed:** Optimized build size.